Flask_API_Server/requirements.txt
Flask==2.3.2
watchdog==3.0.0
Pillow==10.0.1
numpy==1.24.3